requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
reportlab>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
